frontend/course: extract assignments student list, and convert to React.FC
